Added retrieval and saving of sales requisition items from database

+++ src/main/java/com/pj/magic/dao/ProductDao.java
@@ -10,6 +10,6 @@ public interface ProductDao {
 
 	Product findProductByCode(String code);
 
-	Product getProduct(int id);
+	Product getProduct(long id);
 	
 }
