+++ src/main/java/com/pj/magic/service/ProductService.java
@@ -10,6 +10,6 @@ public interface ProductService {
 	
 	Product findProductByCode(String code);
 	
-	Product getProduct(int id);
+	Product getProduct(long id);
 	
 }
